% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_cor_of_cor}
\alias{calc_cor_of_cor}
\title{Computing cor of cor between study expression data}
\usage{
calc_cor_of_cor(ex_list)
}
\arguments{
\item{ex_list}{expression datasets}
}
\value{
cor_of_cors
}
\description{
Computing cor of cor between study expression data
}
