% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_installed}
\alias{check_installed}
\title{Function for checking needed packages are installed}
\usage{
check_installed(packages_needed)
}
\arguments{
\item{packages_needed}{vector of packages to check}
}
\value{
nothing, but will throw error if packages not installed
}
\description{
Function for checking needed packages are installed
}
