% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_construct.R
\name{compare_networks}
\alias{compare_networks}
\title{compare_networks}
\usage{
compare_networks(
  net_membership_1,
  net_membership_2,
  K = 75,
  memb_cut = 0.5,
  na_flag = c("none", "both", "either")
)
}
\arguments{
\item{net_membership_1}{net_membership_1}

\item{net_membership_2}{net_membership_2}

\item{K}{K}

\item{memb_cut}{memb_cut}

\item{na_flag}{na_flag}
}
\value{
data.frame: adjRand, cor_of_cor, cor_coph, overlap
}
\description{
compare_networks
}
\examples{
\dontrun{
compare_networks(
  GSE39582_icwgcna$community_membership,
  read_icwgcna$community_membership
)
}
}
