% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_2study_rbh_Correlation_Based}
\alias{compute_2study_rbh_Correlation_Based}
\title{Compute RBH for Two Metagene Datasets}
\usage{
compute_2study_rbh_Correlation_Based(
  meta_g1,
  meta_g2,
  lower_quant = 0,
  upper_quant = 1,
  max_rank = 1,
  abs = FALSE,
  sparse = TRUE,
  method = "pearson"
)
}
\arguments{
\item{meta_g1, meta_g2}{metagene matrices from two different datasets. Row
names of these matrices must be gene names, and the column names should be
unique metagene names.}

\item{lower_quant}{indicates the quantile for the minimum correlation
for the reciprocal best hits we will find.}

\item{upper_quant}{indicates the quantile for correlations above which
ANY metagene pairing will be considered a "hit."}

\item{max_rank}{represents highest column and row rankings accepted for our
reciprocal best hits network. Pure reciprocal best hits uses max_rank of 1.}

\item{abs}{logical, take absolute values of correlations?}

\item{sparse}{logical, use a sparse matrix to store network?}

\item{method}{string, same as \code{\link[stats:cor]{stats::cor()}}}
}
\value{
Matrix with rows matching the columns in meta_g1 and columns
matching the columns of meta_g2, with values 1 indicating two metagenes are a
hit, and 0 indicating that they are not.
}
\description{
Computes a single network with two datasets. They must have rownames
containing some sort of common geneID vocabulary, so that they can
be intersected to get common rows.
}
\details{
If one wants the reciprocal best hits network to have separate dimension
names for rows and columns, the column names for meta_g1 and meta_g2 should
be distinct.
}
\examples{
\dontrun{
network_file_dir <- system.file("extdata", package = "consensusNetR")
network_file_list <- list.files(network_file_dir, full.names = TRUE)
ds1 <- as.matrix(data.table::fread(network_file_list[[1]]))
row.names(ds1) <- ds1[, 1]
ds1 <- ds1[, -1]
colnames(ds1) <- paste0(colnames(ds1), sub(
  ".*/", "",
  network_file_list[[1]]
))

ds2 <- as.matrix(data.table::fread(network_file_list[[2]]))
row.names(ds2) <- ds2[, 1]
ds2 <- ds2[, -1]
colnames(ds2) <- paste0(colnames(ds2), sub(
  ".*/", "",
  network_file_list[[2]]
))
rbh <- compute_2study_rbh_Correlation_Based(ds1, ds2)
}
}
