% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_mean_meta_genes}
\alias{compute_mean_meta_genes}
\title{Compute the average metagene across studies for each community}
\usage{
compute_mean_meta_genes(
  consensus,
  rbh,
  network_file_list = list(),
  network_file_dir = NULL,
  weights = NULL,
  compressIntra = "first"
)
}
\arguments{
\item{rbh}{matrix indicating edges of recipracol best hit network}

\item{network_file_list}{a list of paths to files which can be read by
data.table::fread. These constitute the metagene datasets. Each file should
have some sort of common gene_id as the first column}

\item{network_file_dir}{a directory in which all files are relevant to the
meta reciprocal best hits, all readable by fread. These will be added to the
file list. Each file should have some sort of common gene_id as the first
column.}

\item{weights}{weights for weighted average of based on study attributes,
not yet implemented}

\item{compressIntra}{indicates how to deal with multiple metagenes beloning
to the same comminity within one study-level network}

\item{comms}{communities output from MCL::mcl, a list containing element
$Cluster with a vector of the cluster associated with each index.}
}
\value{
matrix with average loadings (metagenes) across studies included from
network_file_list
}
\description{
Compute the average metagene across studies for each community
}
\examples{
network_file_dir <- system.file("extdata", package = "consensusNet")
network_file_list <- list.files(network_file_dir, full.names = TRUE)
ma <- construct_meta_rbh(
  network_file_list = network_file_list,
  upper_quant = .99,
  lower_quant = .05, max_rank = 2
)
comms <- detect_metagene_communities(ma)
compute_mean_meta_genes <- function(commms,ma,network_file_list = network_file_list)
}
