% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbh_2study.R
\name{construct_2study_rbh_correlation_based}
\alias{construct_2study_rbh_correlation_based}
\title{Compute correlation based RBH for Two Metagene Datasets}
\usage{
construct_2study_rbh_correlation_based(
  net_membership_1,
  net_membership_2,
  lower_quant = 0,
  upper_quant = 1,
  max_rank = 1,
  abs = FALSE,
  sparse = FALSE,
  method = "pearson"
)
}
\arguments{
\item{net_membership_1}{community membership matrix from first network.}

\item{net_membership_2}{community membership matrix from second network.}

\item{lower_quant}{indicates the quantile for the minimum correlation
for the reciprocal best hits we will find.}

\item{upper_quant}{indicates the quantile for correlations above which
ANY metagene pairing will be considered a "hit."}

\item{max_rank}{represents highest column and row rankings accepted for our
reciprocal best hits network. Pure reciprocal best hits uses max_rank of 1.}

\item{abs}{logical, take absolute values of correlations?}

\item{sparse}{logical, use a sparse matrix to store network?}

\item{method}{string, same as \code{\link[stats:cor]{stats::cor()}}}
}
\value{
Matrix with rows matching the columns in net_membership_1 and columns
matching the columns of net_membership_2, with values 1 indicating two metagenes are a
hit, and 0 indicating that they are not.
}
\description{
Computes a single network with two datasets. They must have rownames
containing some sort of common geneID vocabulary, so that they can
be intersected to get common rows.
}
\details{
If one wants the reciprocal best hits network to have separate dimension
names for rows and columns, the column names for net_membership_1 and net_membership_2 should
be distinct.
}
\examples{
\dontrun{
rbh <- construct_2study_rbh_correlation_based(
  GSE39582_icwgcna$community_membership,
  read_icwgcna$community_membership
)
}
}
