% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_mulit_rbh_overlap_based}
\alias{construct_mulit_rbh_overlap_based}
\title{Construct Meta Reciprocal Best Hits bases on overlaps}
\usage{
construct_mulit_rbh_overlap_based(
  network_membership_list,
  top_n = 50,
  memb_cut = 0
)
}
\arguments{
\item{network_membership_list}{a list containing community membership scores for each
network. Where rownames contain unique gene ids and column names are commuity names}

\item{top_n}{the number of top genes based on membership scores with higher scores
indicating stronger membership in a community}

\item{memb_cut}{membership threshold for stricter thresholding. Only genes with
membership score greater the threshold are used}
}
\description{
Iterate through a list containing membership matrices and construct pairwise
networks for every combination, and output them as a symmetric matrix with
labeled dimensions. This can be interpreted as an "adjacency" matrix between
all of the different metagenes across networks in the input list. The
column/row names will take each networks communities' names and concatenate them
with the network list names, such that the dimension names are guaranteed to
be unique.
}
