% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_metagene_communities}
\alias{detect_metagene_communities}
\title{Detect Communities in Adjacency/Reciprocal Best Hits Matrix}
\usage{
detect_metagene_communities(rbh_mat, expansion = 2)
}
\arguments{
\item{rbh_mat}{adjacency/reciprocal best hits matrix, as output by
\code{\link[=construct_multi_rbh_correlation_based]{construct_multi_rbh_correlation_based()}}}

\item{expansion}{expansion parameter}
}
\value{
The output of \code{\link[MCL:mcl]{MCL::mcl()}}, renamed so that the clusters
are in order of size (1 being the largest)
}
\description{
Performs Markov Clustering on the adjacency matrix, giving the
cluster/community number that each metagene is associated with, renamed so
that the clusters are in order of size (1 being the largest)
}
\examples{
\dontrun{
network_file_dir <- system.file("extdata", package = "consensusNetR")
network_file_list <- list.files(network_file_dir, full.names = TRUE)
ma <- construct_multi_rbh_correlation_based(
  network_file_list = network_file_list,
  upper_quant = .99,``
  lower_quant = .05, max_rank = 2
)
comms <- detect_metagene_communities(ma)
}
}
