% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_consensus_eig_dist}
\alias{plot_consensus_eig_dist}
\title{Plot individual eigengene distributions}
\usage{
plot_consensus_eig_dist(
  eigen_list,
  target_study_index = 1,
  filename = NA,
  width = 12,
  height = 10,
  dpi = 1000
)
}
\arguments{
\item{eigen_list}{eigen_list}

\item{target_study_index}{target_study_index}

\item{filename}{File name}

\item{width}{figure width}

\item{height}{figure height}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}
}
\value{
eigengene distributions if \code{filename} is na, and saved figure is
\code{filename} provided
}
\description{
Plot individual eigengene distributions
}
