% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_rbh}
\alias{plot_rbh}
\title{RBH Heatmap Creation}
\usage{
plot_rbh(rbh, memb_list, file_name = NA, width = 10, height = 8)
}
\arguments{
\item{rbh}{reciprocal best hits matrix (output of
\code{\link[=construct_multi_rbh_correlation_based]{construct_multi_rbh_correlation_based()}} or
\code{\link[=construct_multi_rbh_overlap_based]{construct_multi_rbh_overlap_based()}})}

\item{memb_list}{memb_list?}

\item{file_name}{File name}

\item{width}{figure width}

\item{height}{figure height}
}
\value{
pheatmap object
}
\description{
RBH Heatmap Creation
}
