% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_membership.R
\name{calc_consensus_memberships}
\alias{calc_consensus_memberships}
\title{Compute the average metagene across studies for each community}
\usage{
calc_consensus_memberships(
  consensus_comms,
  network_membership_list,
  weights = NULL,
  gene_cohort_N = 3,
  compressIntra = c("first", "mean"),
  remove_misc_comm = TRUE,
  comm_prefix = "mA"
)
}
\arguments{
\item{consensus_comms}{communities output from \code{\link[MCL:mcl]{MCL::mcl()}},
a list containing elements \code{Cluster} with a vector of the cluster
associated with each index.}

\item{network_membership_list}{a list containing community membership scores for each
network. Where rownames contain unique gene ids and column names are community names}

\item{weights}{weights for weighted average of based on study attributes}

\item{gene_cohort_N}{gene_cohort_N}

\item{compressIntra}{indicates how to deal with multiple metagenes belonging
to the same community within one study-level network}

\item{remove_misc_comm}{should the miscellaneous community be removed
(1st community)}

\item{comm_prefix}{the prefix to add to community names}
}
\value{
matrix with average loadings (metagenes) across studies included from
network_membership_list
}
\description{
Compute the average metagene across studies for each community
}
\examples{
\dontrun{
memb_list <- list(
  GSE39582 = GSE39582_icwgcna$community_membership,
  READ = read_icwgcna$community_membership,
  COAD = coad_icwgcna$community_membership
)
ma <- construct_rbh_correlation_based(
  network_membership_list,
  upper_quant = .99,
  lower_quant = .05,
  max_rank = 2
)
consensus_comms <- detect_consensus_communities(ma)
calc_consensus_memberships(consensus_comms,memb_list)
}
}
