% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_construct.R
\name{calc_cor_of_cor}
\alias{calc_cor_of_cor}
\title{Computing cor of cor between study expression data}
\usage{
calc_cor_of_cor(ex_list)
}
\arguments{
\item{ex_list}{list of expression datasets}
}
\value{
correlations of correlations
}
\description{
Computing cor of cor between study expression data
}
\examples{
\dontrun{
ex_list <- list(
  GSE39582 = GSE39582_df,
  READ = read_df,
  COAD = coad_df
)

calc_cor_of_cor(ex_list)
}
}
