% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_mean_meta_genes}
\alias{compute_mean_meta_genes}
\title{Compute the average metagene across studies for each community}
\usage{
compute_mean_meta_genes(
  consensus,
  network_membership_list,
  weights = NULL,
  gene_cohort_N = 3,
  compressIntra = "first"
)
}
\arguments{
\item{consensus}{communities output from \code{\link[MCL:mcl]{MCL::mcl()}}, a list containing element
$Cluster with a vector of the cluster associated with each index.}

\item{network_membership_list}{a list containing community membership scores for each
network. Where rownames contain unique gene ids and column names are community names}

\item{weights}{weights for weighted average of based on study attributes,
not yet implemented}

\item{gene_cohort_N}{gene_cohort_N}

\item{compressIntra}{indicates how to deal with multiple metagenes belonging
to the same community within one study-level network}
}
\value{
matrix with average loadings (metagenes) across studies included from
network_file_list
}
\description{
Compute the average metagene across studies for each community
}
\examples{
\dontrun{
network_file_dir <- system.file("extdata", package = "consensusNetR")
network_file_list <- list.files(network_file_dir, full.names = TRUE)
ma <- construct_meta_rbh(
  network_file_list = network_file_list,
  upper_quant = .99,
  lower_quant = .05, max_rank = 2
)
comms <- detect_metagene_communities(ma)
compute_mean_meta_genes(comms,ma,network_file_list = network_file_list)
}
}
