% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbh_2study.R
\name{construct_2study_rbh_overlap_based}
\alias{construct_2study_rbh_overlap_based}
\title{Compute RBH (reciprocal best hits) for Two Networks based on Overlap of top genes}
\usage{
construct_2study_rbh_overlap_based(meta_g1, meta_g2, top_n = 50, memb_cut = 0)
}
\arguments{
\item{meta_g1}{community membership matrix from first network.}

\item{meta_g2}{community membership matrix from second network.
Row names of these matrices must be gene names, and the column names should be
unique community names.}

\item{top_n}{number of top genes to usin in compuitng overlap}

\item{memb_cut}{membership threshold for stricter thresholding. Only genes with
membership score greater the threshold are used}
}
\value{
Matrix with rows matching the columns in meta_g1 and columns
matching the columns of meta_g2, with values 1 indicating two metagenes are a
hit, and 0 indicating that they are not.
}
\description{
Compute RBH (reciprocal best hits) for Two Networks based on Overlap of top
community genes. They must have row names containing some sort of common geneID
vocabulary, so that they can be intersected to get common rows.
}
