% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_unique_top_genes}
\alias{find_unique_top_genes}
\title{Identify Top Gene of Communities that are unique (only belong to one community)}
\usage{
find_unique_top_genes(mGenes, K = 10, max_iter = 10)
}
\arguments{
\item{mGenes}{meta gene matrix rownames being gene IDs and columns being
community loadings (i.e. metagenes).}

\item{K}{number of unique genes to find.}

\item{max_iter}{maximum number of iterations to avoid looping to meaningless top genes}
}
\value{
data.frame with key = community and value = top gene ID. Top genes will only
be associated with one module.
}
\description{
Identify Top Gene of Communities that are unique (only belong to one community)
}
