% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gene_community_membership}
\alias{get_gene_community_membership}
\title{Get Consensus Membership of Genes in Metagene Communities}
\usage{
get_gene_community_membership(
  comms,
  min_studies,
  network_file_list = list(),
  network_file_dir = NULL,
  include_nonmembers = FALSE,
  compress = TRUE,
  rank_based = TRUE
)
}
\arguments{
\item{comms}{communities output from MCL::mcl, a list containing element
$Cluster with a vector of the cluster associated with each index.}

\item{min_studies}{the minimum threshold of studies/datasets indicating a
gene belongs to a community. In a given dataset, a gene "belongs" to a
community if the gene's highest loading corresponds to that community.}

\item{network_file_list}{a list of paths to files which can be read by
data.table::fread. These constitute the metagene datasets. Each file should
have some sort of common gene_id as the first column}

\item{network_file_dir}{a directory in which all files are relevant to the
meta reciprocal best hits, all readable by fread. These will be added to the
file list. Each file should have some sort of common gene_id as the first
column.}

\item{include_nonmembers}{logical indicating whether or not to include genes
which do not appear in a community.}

\item{compress}{indicates whether to drop duplicate community metagenes. Assumes
first one is the one to keep (largest or best). Currently only TRUE is supported}

\item{rank_based}{flag indicating whether to use rank when determing max metagene.
Currently only TRUE is supported}
}
\value{
data.frame containing gene ids and their associated most likely
cluster(s) in $cluster and $n_studies, representing the number of studies
or datasets indicating a gene belongs to the most associated community. In a
given dataset, a gene "belongs" to a community if the gene's highest loading
corresponds to that community.
}
\description{
Get Consensus Membership of Genes in Metagene Communities
}
\examples{
\dontrun{
network_file_dir <- system.file("extdata", package = "consensusNetR")
network_file_list <- list.files(network_file_dir, full.names = TRUE)
ma <- construct_meta_rbh(
  network_file_list = network_file_list,
  upper_quant = .99,
  lower_quant = .05, max_rank = 2
)
comms <- detect_metagene_communities(ma)
consensus <- get_gene_community_membership(comms, 2,
  network_file_dir = network_file_dir
)
}
}
