% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbh.R
\name{plot_rbh}
\alias{plot_rbh}
\title{RBH Heatmap Creation}
\usage{
plot_rbh(rbh, network_membership_list, file_name = NA, width = 10, height = 8)
}
\arguments{
\item{rbh}{reciprocal best hits matrix (output of
\code{\link[=construct_rbh_correlation_based]{construct_rbh_correlation_based()}} or
\code{\link[=construct_rbh_overlap_based]{construct_rbh_overlap_based()}})}

\item{network_membership_list}{a list containing community membership scores for each
network. Where rownames contain unique gene ids and column names are community names}

\item{file_name}{File name}

\item{width}{figure width}

\item{height}{figure height}
}
\value{
pheatmap object
}
\description{
RBH Heatmap Creation
}
\examples{
\dontrun{
memb_list <- list(
  GSE39582 = GSE39582_icwgcna$community_membership,
  READ = read_icwgcna$community_membership,
  COAD = coad_icwgcna$community_membership
)
ma <- construct_rbh_overlap_based(memb_list)
plot_rbh(ma, memb_list)
}
}
